Added clear and tested installation instructions
The installation instructions as they were were not very clear to me. I had to do many things that were not written anywhere in the documents or online, so I had to figure it out as I went and wanted to give back and help other people use this neat piece of software.